fix: overflow for big N (#23)

+++ omp/simulation.c
@@ -1,14 +1,14 @@
 #include "simulation.h"
 
 unsigned char *grid;
-int gridSize;
-int gridPadding;
+long long gridSize;
+long long gridPadding;
 unsigned char *auxState;
 int genNum;
 
 long long leaderboard[(N_SPECIES + 1) * 3] = {0}; // current, max, max gen
 
-void initializeAux(unsigned char *g, int num, int size) {
+void initializeAux(unsigned char *g, int num, long long size) {
   grid = g;
   gridSize = size;
   gridPadding = size + 2;
@@ -27,11 +27,11 @@ void simulation() {
 #pragma omp parallel
   {
 #pragma omp for reduction(+ : leaderboard[ : N_SPECIES + 1])
-    for (int z = 1; z < gridPadding - 1; z++) {
-      int z_ = z * gridPadding * gridPadding;
-      for (int y = 1; y < gridPadding - 1; y++) {
-        int y_ = y * gridPadding;
-        for (int x = 1; x < gridPadding - 1; x++) {
+    for (long long z = 1; z < gridPadding - 1; z++) {
+      long long z_ = z * gridPadding * gridPadding;
+      for (long long y = 1; y < gridPadding - 1; y++) {
+        long long y_ = y * gridPadding;
+        for (long long x = 1; x < gridPadding - 1; x++) {
           leaderboard[grid[z_ + y_ + x]]++;
         }
       }
@@ -44,11 +44,11 @@ void simulation() {
     for (int gen = 1; gen < genNum + 1; gen++) {
 
 #pragma omp for reduction(+ : leaderboard[ : N_SPECIES + 1])
-      for (int z = 1; z < gridPadding - 1; z++) {
-        int z_ = z * gridPadding * gridPadding;
-        for (int y = 1; y < gridPadding - 1; y++) {
-          int y_ = y * gridPadding;
-          for (int x = 1; x < gridPadding - 1; x++) {
+      for (long long z = 1; z < gridPadding - 1; z++) {
+        long long z_ = z * gridPadding * gridPadding;
+        for (long long y = 1; y < gridPadding - 1; y++) {
+          long long y_ = y * gridPadding;
+          for (long long x = 1; x < gridPadding - 1; x++) {
             leaderboard[updateCellState(x, y, z, z_ + y_ + x)]++;
           }
         }
@@ -69,7 +69,8 @@ void simulation() {
   }
 };
 
-unsigned char updateCellState(int x, int y, int z, int index) {
+unsigned char updateCellState(long long x, long long y, long long z,
+                              long long index) {
 
   unsigned char current_state = grid[index];
 
@@ -82,25 +83,25 @@ unsigned char updateCellState(int x, int y, int z, int index) {
   return new_state;
 };
 
-void writeCellState(int x, int y, int z, int index, unsigned char old_state,
-                    unsigned char value) {
+void writeCellState(long long x, long long y, long long z, long long index,
+                    unsigned char old_state, unsigned char value) {
 
   auxState[index] = value;
   writeBorders(auxState, gridPadding, x, y, z, value);
 };
 
-unsigned char getNeighbourCount(int x, int y, int z) {
+unsigned char getNeighbourCount(long long x, long long y, long long z) {
   unsigned char count = 0;
 
-  int z1 = (z - 1) * gridPadding * gridPadding;
-  int z2 = z * gridPadding * gridPadding;
-  int z3 = (z + 1) * gridPadding * gridPadding;
-  int y1 = (y - 1) * gridPadding;
-  int y2 = y * gridPadding;
-  int y3 = (y + 1) * gridPadding;
-  int x1 = x - 1;
-  int x2 = x;
-  int x3 = x + 1;
+  long long z1 = (z - 1) * gridPadding * gridPadding;
+  long long z2 = z * gridPadding * gridPadding;
+  long long z3 = (z + 1) * gridPadding * gridPadding;
+  long long y1 = (y - 1) * gridPadding;
+  long long y2 = y * gridPadding;
+  long long y3 = (y + 1) * gridPadding;
+  long long x1 = x - 1;
+  long long x2 = x;
+  long long x3 = x + 1;
 
   count += (grid[z1 + y1 + x1] != 0);
   count += (grid[z1 + y1 + x2] != 0);
@@ -135,8 +136,8 @@ unsigned char getNeighbourCount(int x, int y, int z) {
 };
 
 // wraps around the grid
-unsigned char calculateNextState(int x, int y, int z,
-                                 unsigned char current_state, int index) {
+unsigned char calculateNextState(long long x, long long y, long long z,
+                                 unsigned char current_state, long long index) {
 
   unsigned char neighbourCount = getNeighbourCount(x, y, z);
   if (current_state == 0) {
@@ -150,18 +151,18 @@ unsigned char calculateNextState(int x, int y, int z,
   return (neighbourCount <= 4 || neighbourCount > 13) ? 0 : current_state;
 };
 
-unsigned char getMostFrequentValue(int x, int y, int z) {
+unsigned char getMostFrequentValue(long long x, long long y, long long z) {
   unsigned char neighborsValues[N_SPECIES + 1] = {0};
 
-  int z1 = (z - 1) * gridPadding * gridPadding;
-  int z2 = z * gridPadding * gridPadding;
-  int z3 = (z + 1) * gridPadding * gridPadding;
-  int y1 = (y - 1) * gridPadding;
-  int y2 = y * gridPadding;
-  int y3 = (y + 1) * gridPadding;
-  int x1 = x - 1;
-  int x2 = x;
-  int x3 = x + 1;
+  long long z1 = (z - 1) * gridPadding * gridPadding;
+  long long z2 = z * gridPadding * gridPadding;
+  long long z3 = (z + 1) * gridPadding * gridPadding;
+  long long y1 = (y - 1) * gridPadding;
+  long long y2 = y * gridPadding;
+  long long y3 = (y + 1) * gridPadding;
+  long long x1 = x - 1;
+  long long x2 = x;
+  long long x3 = x + 1;
 
   neighborsValues[grid[z1 + y1 + x1]]++;
   neighborsValues[grid[z1 + y1 + x2]]++;
@@ -204,10 +205,10 @@ unsigned char getMostFrequentValue(int x, int y, int z) {
 };
 
 void debugPrintGrid() {
-  for (int z = 0; z < gridPadding; z++) {
-    for (int y = 0; y < gridPadding; y++) {
-      for (int x = 0; x < gridPadding; x++) {
-        int index = z * gridPadding * gridPadding + y * gridPadding + x;
+  for (long long z = 0; z < gridPadding; z++) {
+    for (long long y = 0; y < gridPadding; y++) {
+      for (long long x = 0; x < gridPadding; x++) {
+        long long index = z * gridPadding * gridPadding + y * gridPadding + x;
         int valueToPrint = (int)grid[index];
         if (valueToPrint == 0) {
           fprintf(stdout, "  ");
